+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Queue daemon which can run the queue reliably twice each minute, similar to the cron daemon.
 - New method to generate a random filename with 25 characters: `C::Storage()->getRandomFilename()`.
 - New method `C::getAllAppModules()` to only get app-related loaded modules like the App module, no kernel modules.
+- Slugifying is improved and can now create safe filenames via `C::Formatter()->slugify($text, true)`.
 
 ### 🔧 Changed